backend/pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "albertoa", email = "aabarzua@palvi.cl"},
]
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14.0",
    "psycopg2-binary>=2.9.6",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pandas>=2.0.1",
    "twelvedata>=1.2.11",
    "setuptools>=67.7.2",
    "pymemcache>=4.0.0",
    "flake8-print>=5.0.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}


[tool.isort]
profile = "black"

[tool.black]
line-length = 90
[tool.pdm.scripts]
isort = "isort . --skip .venv"
black = "black . --exclude .venv"
lint = "flake8 ."
format = {composite = ["isort", "black"]}
fixtures = "pdm run python manage.py loaddata fixtures/*.jsonpdm run py"
test = "pdm run python manage.py test --pattern=\"test_*.py\""
makemigrations = "pdm run python manage.py makemigrations --noinput"
migrate = "pdm run python manage.py migrate --noinput"
mm = {composite = ["makemigrations", "migrate"]}
